+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [2.12.0] - 2024-06-20Z
 - Fix Student Loan rounding issue, and added new parameterized test for Student Loan.
 - Fix Scottish Tax Rate by adding allowing SD3 tax code for Scottish top rate.
 - Added Clarification for Pension Allowance, and adjusted Pension Allowance amount.